extract Megatron Mapper for VLM
